Update directory, tests and requirements
