--- python/regex_patterns/data/companies.json
@@ -1,24 +1,0 @@
-{
-    "companies": [
-        {
-            "company": "Acme Inc.",
-            "annual_revenue": "$10M-$20M",
-            "industry": "Technology"
-        },
-        {
-            "company": "Globex Corp.",
-            "annual_revenue": "$50M-$100M",
-            "industry": "Logistics"
-        },
-        {
-            "company": "Initech Corp.",
-            "annual_revenue": "$500K-$1M",
-            "industry": "Technology"
-        },
-        {
-            "company": "Umbrella Corp.",
-            "annual_revenue": "$800M-$1B",
-            "industry": "Retail"
-        }
-    ]
-}
